requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
Faker==24.3.0
frozenlist==1.4.1
gino==1.0.1
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
pillow==10.2.0
pydantic==2.5.3
pydantic-settings==2.2.1
pydantic_core==2.14.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==1.3.24
typing_extensions==4.10.0
urllib3==2.2.1
yarl==1.9.4
